requirements-dev.txt
ansible_core>=2.12.0

# Development requirements
antsibull-docs>=2.24,<2.25
pylint
pytest
pytest-xdist